require phpunit with composer, delete composer.lock
composer lock should not be committed because when devs use composer
to install the sdk the lock file is not used to install our dependencies.
while commiting the lock file makes testing easier and keeps a consistent dev
environment, it is not consistent with the experience  a consumer of the sdk gets.